requirements.txt
tweepy==3.10.0
pandas==1.2.4
requests==2.25.1
python-dotenv==0.17.1
